# Backendbooking/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Space Booking API
  version: 1.0.0
  description: |
    Kontrak 50 operasi UKK Paket B. Semua data bisnis tenant-scoped.
    x-maker-key dan x-app-key adalah alias; jika keduanya hadir nilainya harus sama.
    Kontrak ini sengaja memperketat upload menjadi bearer-only, melindungi maker/list,
    menggunakan QR opaque tanpa app key, dan mempertahankan body array collection dengan
    metadata pagination pada X-Page/X-Per-Page/X-Total-Count. Lihat api-contract.md untuk conflict log.
servers:
  - url: https://learn.smktelkom-mlg.sch.id/coworking
    description: Base URL contoh sumber UKK
  - url: http://localhost:3000
    description: Development lokal
...
paths:
...
  /api/spaces/types:
    get:
      tags: [Spaces]
      operationId: listSpaceTypes
      summary: Daftar tipe space
      security:
        - makerKey: []
        - appKey: []
      responses:
        '200':
          description: Tiga tipe space
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceTypeListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/spaces/availability:
    get:
      tags: [Spaces]
      operationId: checkSpaceAvailability
      summary: Cek availability advisory
      security:
        - makerKey: []
        - appKey: []
      parameters:
        - name: id_space
          in: query
          required: true
          schema: { $ref: '#/components/schemas/PositiveId' }
        - name: tanggal
          in: query
          required: true
          schema: { $ref: '#/components/schemas/DateOnly' }
        - name: jam_mulai
          in: query
          required: true
          schema: { $ref: '#/components/schemas/TimeHHmm' }
        - name: durasi_jam
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Hasil availability
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/spaces:
    get:
      tags: [Spaces]
      operationId: listPublicSpaces
      summary: Katalog space tenant
      security:
        - makerKey: []
        - appKey: []
      parameters:
        - name: tipe
          in: query
          schema: { $ref: '#/components/schemas/SpaceType' }
        - name: search
          in: query
          schema: { type: string, minLength: 1, maxLength: 100 }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Daftar space
          headers:
            X-Page: { $ref: '#/components/headers/XPage' }
            X-Per-Page: { $ref: '#/components/headers/XPerPage' }
            X-Total-Count: { $ref: '#/components/headers/XTotalCount' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/spaces/{id}:
    get:
      tags: [Spaces]
      operationId: getPublicSpaceById
      summary: Detail space
      security:
        - makerKey: []
        - appKey: []
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200':
          description: Detail space
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/admin/spaces:
    get:
      tags: [Admin Spaces]
      operationId: listAdminSpaces
      summary: Daftar space milik admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Daftar space admin
          headers:
            X-Page: { $ref: '#/components/headers/XPage' }
            X-Per-Page: { $ref: '#/components/headers/XPerPage' }
            X-Total-Count: { $ref: '#/components/headers/XTotalCount' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Admin Spaces]
      operationId: createAdminSpace
      summary: Tambah space
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSpaceDto' }
      responses:
        '201':
          description: Space dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/admin/spaces/{id}:
    parameters:
      - { $ref: '#/components/parameters/IdPath' }
    get:
      tags: [Admin Spaces]
      operationId: getAdminSpaceById
      summary: Detail space milik admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      responses:
        '200':
          description: Detail space
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [Admin Spaces]
      operationId: updateAdminSpace
      summary: Update space milik admin
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSpaceDto' }
      responses:
        '200':
          description: Space diperbarui
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpaceResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Admin Spaces]
      operationId: deleteAdminSpace
      summary: Arsipkan/hapus space tanpa merusak histori
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      responses:
        '200':
          description: Space diarsipkan atau dihapus
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/upload/image:
    post:
      tags: [Uploads]
      operationId: uploadGeneralImage
      summary: Upload gambar umum
      description: Konflik sumber diselesaikan dengan bearer wajib untuk member/admin.
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ImageUploadDto' }
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '201':
          description: File tersimpan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GeneralUploadResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/upload/spaces:
    post:
      tags: [Uploads]
      operationId: uploadSpaceImage
      summary: Upload foto space oleh admin
      description: Konflik sumber diselesaikan dengan bearer admin wajib.
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ImageUploadDto' }
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '201':
          description: Foto space tersimpan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimpleUploadResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/upload/members:
    post:
      tags: [Uploads]
      operationId: uploadMemberImage
      summary: Upload foto member oleh member/admin
      description: Konflik sumber diselesaikan dengan bearer wajib dan ownership saat foto di-attach.
      security:
        - bearerAuth: []
          makerKey: []
        - bearerAuth: []
          appKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ImageUploadDto' }
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        '201':
          description: Foto member tersimpan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimpleUploadResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT App Maker, member, admin_space, atau Guru/Penguji sesuai endpoint.
    makerKey:
      type: apiKey
      in: header
      name: x-maker-key
      description: Tenant key utama.
    appKey:
      type: apiKey
      in: header
      name: x-app-key
      description: Alias kompatibilitas untuk x-maker-key; bila keduanya ada harus sama.
  headers:
    XPage:
      description: Nomor halaman efektif (1-based).
      schema: { type: integer, minimum: 1 }
    XPerPage:
      description: Batas efektif item per halaman.
      schema: { type: integer, minimum: 1, maximum: 100 }
    XTotalCount:
      description: Jumlah total item yang cocok sebelum pagination.
      schema: { type: integer, minimum: 0 }
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/PositiveId' }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: Input/header tidak valid atau aturan bisnis dasar gagal
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { status: false, statusCode: 400, message: Permintaan tidak valid, error: Bad Request, timestamp: '2026-08-27T08:00:00.000Z' }
    Unauthorized:
      description: Credential hilang atau invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Role/ownership/state tidak mengizinkan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource tidak ada atau disamarkan karena scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Duplicate, overlap, stale write, atau transisi ilegal
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    PayloadTooLarge:
      description: File melebihi 5 MiB
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TooManyRequests:
      description: Rate limit terlampaui
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
          description: Detik sebelum mencoba kembali
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Kegagalan internal tanpa stack/SQL/secret
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    PositiveId:
      type: integer
      format: int64
      minimum: 1
      example: 1
    Money:
      type: integer
      format: int64
      minimum: 0
      description: Nominal integer Rupiah (IDR)
      example: 25000
    DateOnly:
      type: string
      format: date
      example: '2026-08-30'
    TimeHHmm:
      type: string
      pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
      example: '09:00'
    SpaceType:
      type: string
      enum: [desk, meeting_room, private_office]
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [status, statusCode, message, error, timestamp]
      properties:
        status: { type: boolean, enum: [false] }
        statusCode: { type: integer, enum: [400, 401, 403, 404, 409, 413, 429, 500] }
        message: { type: string }
        error: { type: string }
        timestamp: { type: string, format: date-time }
...
    SuccessBase:
      type: object
      required: [status, statusCode, message, timestamp]
      properties:
        status: { type: boolean, enum: [true] }
        statusCode: { type: integer, enum: [200, 201] }
        message: { type: string }
        timestamp: { type: string, format: date-time }
...
    SpaceTypeItem:
      type: object
      additionalProperties: false
      required: [tipe, label, deskripsi]
      properties:
        tipe: { $ref: '#/components/schemas/SpaceType' }
        label: { type: string }
        deskripsi: { type: string }
    SpaceTypeListResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data:
              type: array
              minItems: 3
              maxItems: 3
              items: { $ref: '#/components/schemas/SpaceTypeItem' }
    OwnerSummary:
      type: object
      additionalProperties: false
      required: [nama_coworking, nama_pemilik, telp]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        nama_coworking: { type: string }
        nama_pemilik: { type: string }
        telp: { type: string }
    Space:
      type: object
      additionalProperties: false
      required: [id, nama_space, harga_per_jam, tipe, kapasitas, deskripsi, id_owner]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        nama_space: { type: string }
        harga_per_jam: { $ref: '#/components/schemas/Money' }
        tipe: { $ref: '#/components/schemas/SpaceType' }
        kapasitas: { type: integer, minimum: 1 }
        foto: { type: string, nullable: true }
        deskripsi: { type: string }
        id_owner: { $ref: '#/components/schemas/PositiveId' }
        owner: { $ref: '#/components/schemas/OwnerSummary' }
        foto_url: { type: string, format: uri, nullable: true }
    SpaceResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/Space' }
    SpaceListResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data:
              type: array
              maxItems: 100
              items: { $ref: '#/components/schemas/Space' }
    Availability:
      type: object
      additionalProperties: false
      required: [available, id_space, nama_space, tanggal, jam_mulai, jam_selesai, durasi_jam, harga_per_jam, estimasi_total]
      properties:
        available: { type: boolean }
        id_space: { $ref: '#/components/schemas/PositiveId' }
        nama_space: { type: string }
        tanggal: { $ref: '#/components/schemas/DateOnly' }
        jam_mulai: { $ref: '#/components/schemas/TimeHHmm' }
        jam_selesai: { $ref: '#/components/schemas/TimeHHmm' }
        durasi_jam: { type: integer, minimum: 1 }
        harga_per_jam: { $ref: '#/components/schemas/Money' }
        estimasi_total: { $ref: '#/components/schemas/Money' }
    AvailabilityResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/Availability' }
...
    DeleteData:
      type: object
      additionalProperties: false
      required: [id, deleted]
      properties:
        id: { $ref: '#/components/schemas/PositiveId' }
        deleted: { type: boolean, enum: [true] }
    DeleteResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/DeleteData' }
    CreateSpaceDto:
      type: object
      additionalProperties: false
      required: [nama_space, harga_per_jam, tipe, kapasitas, deskripsi]
      properties:
        nama_space: { type: string, minLength: 1, maxLength: 200, example: Personal Desk Alpha 01 }
        harga_per_jam: { $ref: '#/components/schemas/Money' }
        tipe: { $ref: '#/components/schemas/SpaceType' }
        kapasitas: { type: integer, minimum: 1, example: 1 }
        deskripsi: { type: string, minLength: 1, maxLength: 4000 }
        foto: { type: string, minLength: 1, maxLength: 512, nullable: true }
    UpdateSpaceDto:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nama_space: { type: string, minLength: 1, maxLength: 200 }
        harga_per_jam: { $ref: '#/components/schemas/Money' }
        tipe: { $ref: '#/components/schemas/SpaceType' }
        kapasitas: { type: integer, minimum: 1 }
        deskripsi: { type: string, minLength: 1, maxLength: 4000 }
        foto: { type: string, minLength: 1, maxLength: 512, nullable: true }
...
    ImageUploadDto:
      type: object
      additionalProperties: false
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Satu JPEG/PNG/WebP maksimum 5 MiB; MIME, extension, dan magic bytes wajib cocok.
    GeneralUploadData:
      type: object
      additionalProperties: false
      required: [filename, original_name, mimetype, size, url]
      properties:
        filename: { type: string, example: 1787799592972-544446318.jpeg }
        original_name: { type: string, example: banner.jpg }
        mimetype: { type: string, enum: [image/jpeg, image/png, image/webp] }
        size: { type: integer, minimum: 1, maximum: 5242880 }
        url: { type: string, format: uri }
    SimpleUploadData:
      type: object
      additionalProperties: false
      required: [filename, url]
      properties:
        filename: { type: string, example: meeting_room_alpha.jpg }
        url: { type: string, format: uri }
    GeneralUploadResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/GeneralUploadData' }
    SimpleUploadResponse:
      allOf:
        - { $ref: '#/components/schemas/SuccessBase' }
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/SimpleUploadData' }